- Corrected a bug in Cell - Made Cell immutable - Modified CellTest

+++ client/src/main/scala/it/cwmp/client/view/game/model/CellView.scala
@@ -44,5 +44,5 @@ object CellView {
     *
     * Maps size to energy
     */
-  val sizingStrategy: SizingStrategy[Cell, Int] = _.energy
+  val sizingStrategy: SizingStrategy[Cell, Int] = _.energy.toInt
 }
